data-management/saga-orchestration/order-service/src/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
cached-property==2.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
conductor-python==1.1.10
dacite==1.9.2
fastapi==0.116.1
fastapi_class==3.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
inflection==0.5.1
mypy_extensions==1.1.0
prometheus_client==0.22.1
pyconductor==0.1.3
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.5
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-orm==1.2.10
starlette==0.47.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
PyMySQL>=1.1.1